Moved interfaces in plugin project
