+++ EideticMemoryOverlay.PluginApi/Buttons/CardButton.cs
@@ -1,4 +1,6 @@
-﻿namespace EideticMemoryOverlay.PluginApi.Buttons {
+﻿using EideticMemoryOverlay.PluginApi.Interfaces;
+
+namespace EideticMemoryOverlay.PluginApi.Buttons {
     public class CardButton : CardImageButton, ICard {
         public CardButton(CardImageButton button) : base(button.CardInfo, button.IsToggled) {
         }
