+++ EideticMemoryOverlay.PluginApi/CardZone.cs
@@ -1,4 +1,5 @@
 ﻿using EideticMemoryOverlay.PluginApi.Buttons;
+using EideticMemoryOverlay.PluginApi.Interfaces;
 using Emo.Common.Enums;
 using Emo.Common.Services;
 using Emo.Common.Utils;
