+++ EideticMemoryOverlay.PluginApi/Interfaces/IAppData.cs
@@ -1,4 +1,4 @@
-﻿namespace EideticMemoryOverlay.PluginApi {
+﻿namespace EideticMemoryOverlay.PluginApi.Interfaces {
     public interface IAppData {
         string Status { get; set; }
     }
