+++ EideticMemoryOverlay.PluginApi/Interfaces/ICard.cs
@@ -1,4 +1,4 @@
-﻿namespace EideticMemoryOverlay.PluginApi {
+﻿namespace EideticMemoryOverlay.PluginApi.Interfaces {
     public interface ICard {
         CardInfo CardInfo { get; }
     }
