+++ EideticMemoryOverlay.PluginApi/Interfaces/ILoggingService.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace EideticMemoryOverlay.PluginApi {
+namespace EideticMemoryOverlay.PluginApi.Interfaces {
     public interface ILoggingService {
         void LogMessage(string message);
