+++ EideticMemoryOverlay.PluginApi/LocalCards/LocalCardsService.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using EideticMemoryOverlay.PluginApi.Interfaces;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
