+++ EideticMemoryOverlay/Data/Game.cs
@@ -1,5 +1,6 @@
 ﻿using EideticMemoryOverlay.PluginApi;
 using EideticMemoryOverlay.PluginApi.Buttons;
+using EideticMemoryOverlay.PluginApi.Interfaces;
 using Emo.Common.Enums;
 using Emo.Common.Services;
 using Emo.Common.Utils;
