+++ EideticMemoryOverlay/Pages/Main/MainViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using EideticMemoryOverlay.PluginApi;
+using EideticMemoryOverlay.PluginApi.Interfaces;
 using Emo.Data;
 using Emo.Services;
 using PageController;
