+++ EideticMemoryOverlay/Pages/Overlay/OverlayCardViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using EideticMemoryOverlay.PluginApi;
+using EideticMemoryOverlay.PluginApi.Interfaces;
 using Emo.Data;
 using PageController;
 using System;
