+++ EideticMemoryOverlay/Pages/SelectCards/SelectCardsViewModel.cs
@@ -1,4 +1,4 @@
-﻿using EideticMemoryOverlay.PluginApi;
+﻿using EideticMemoryOverlay.PluginApi.Interfaces;
 using PageController;
 
 namespace Emo.Pages.SelectCards {
