+++ Plugins/ArkhamHorrorLcg/ArkhamConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using EideticMemoryOverlay.PluginApi;
+using EideticMemoryOverlay.PluginApi.Interfaces;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
