+++ Plugins/ArkhamHorrorLcg/ArkhamDb/ArkhamDbService.cs
@@ -1,4 +1,4 @@
-﻿using EideticMemoryOverlay.PluginApi;
+﻿using EideticMemoryOverlay.PluginApi.Interfaces;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
