+++ Plugins/ArkhamHorrorLcg/ArkhamHorrorLcg.cs
@@ -1,6 +1,7 @@
 ﻿using ArkhamHorrorLcg.ArkhamDb;
 using EideticMemoryOverlay.PluginApi;
 using EideticMemoryOverlay.PluginApi.Buttons;
+using EideticMemoryOverlay.PluginApi.Interfaces;
 using EideticMemoryOverlay.PluginApi.LocalCards;
 using Emo.Common.Enums;
 using StructureMap;
